pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
orjson = "^3.8.6"
tqdm = "^4.64.1"
pyarrow = "^11.0.0"
pylance = "^0.3.16"
ftfy = "^6.1.1"
sparse-dot-topn = "^0.3.4"
scikit-learn = "^1.2.2"
spacy = "^3.5.1"
mgzip = "^0.2.1"
gensim = "^4.3.1"
tensorflow = "^2.12.0"
tensorflow-hub = "^0.13.0"
polyleven = "^0.8"
cython = "^0.29.33"
msgpack-python = "^0.5.6"
progressbar2 = "^4.2.0"
pycapnp = "^1.3.0"
marisa-trie = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
